Hankki_Dining_Platform/fourmen_webapp/requirements.txt
bcrypt==3.1.4
cffi==1.11.5
click==6.7
Flask==0.12.2
Flask-Assets==0.12
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.4
pycparser==2.18
pypostgres==0.8.2
six==1.11.0
SQLAlchemy==1.2.5
webassets==0.12.1
Werkzeug==0.14.1
elasticsearch==5.1.0
minio==3.0.4
pytest==3.6.0
coverage==4.5.1
iamport-rest-client==0.7.0
